fix: some test cases do not require a login (#1253)
![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/6EhJvt3Bpd5hwa82Zuef/8b7c8ac7-b502-4e08-b0fa-1f4888a3fcbb.png)

The `bun test` command is currently experiencing many failures. Upon reviewing the logs, I noticed that a bunch of test cases crash when invoking common code of `backend/src/trpc/base.ts`. This could be because some TRPC interfaces are publicly accessible and were not included in the recent changes.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/6EhJvt3Bpd5hwa82Zuef/50e9ae50-3651-4831-b80b-59676e91642c.png)

This hotfix can resolve the above issue:

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/6EhJvt3Bpd5hwa82Zuef/da3047e4-ad11-436e-8efc-5e514d9b215c.png)